Move update_completion_percentage to webhelpers, since it is called from within BrowserViews
