+++ src/euphorie/client/browser/module.py
@@ -151,7 +151,7 @@ class IdentificationView(BrowserView):
                 context.postponed = False
             else:
                 context.postponed = True
-            self.aq_parent.session.update_completion_percentage()
+            self.webhelpers.update_completion_percentage(self.webhelpers.traversed_session.session)
             self.aq_parent.session.touch()
 
         if reply.get("next") == "previous":
