+++ src/euphorie/client/browser/risk.py
@@ -195,7 +195,7 @@ class IdentificationView(BrowserView):
                         self.calculatePriority(self.risk, reply)
                     elif self.risk is None or self.risk.evaluation_method == "direct":
                         self.context.priority = reply.get("priority")
-                self.session.update_completion_percentage()
+                self.webhelpers.update_completion_percentage(self.session)
 
             if self.use_existing_measures and reply.get("handle_measures_in_place"):
                 measures = self.get_existing_measures()
